server/nodes/skymap/sensor_array/pyproject.toml
[project]
name = "swarm-node"
description = "middleware stack for real time remote operation of connected devices"
version = "0.0.7"
requires-python = "~=3.11.0"
classifiers = [
    "Environment :: No Input/Output (Daemon)",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "uvloop==0.21.0",
    "msgspec~=0.19.0",
    "pyudev==0.24.3",
    "pyserial==3.5",
    "numpy~=2.0",
    "opencv-python>=4.10.0.84",
    "pyrealsense2~=2.55.1",
    "pynmeagps==1.0.43",
    "google-crc32c>=1",
    "httpx>=0.28.1",
    "rp2040",
    "webrtc-proxy",
    "swarmnode-skymap-common",
]

[tool.uv.sources]
rp2040 = { path = "../../../utils/rp2040" }
webrtc-proxy = { path = "../../../webrtc-proxy/grpc/python", editable = true }
swarmnode-skymap-common = { path = "../common", editable = true }

[dependency-groups]
dev = [
    "matplotlib>=3.10.0",
    "pillow>=11.1.0",
]
